Update Python for Data Professional Beginner - Part 3.py
kurang int

+++ Learn/Python/Fundamental/Python for Data Professional Beginner - Part 3/Python for Data Professional Beginner - Part 3.py
@@ -345,7 +345,7 @@ class TenagaLepas(Karyawan):
     def __init__(self, nama, usia, pendapatan):
         super().__init__(nama, usia, pendapatan,0)
     def tambahan_proyek(self, nilai_proyek):
-        self.pendapatan_tambahan += nilai_proyek * 0.01
+        self.pendapatan_tambahan += int(nilai_proyek * 0.01)
 
 
 # # [Tugas Praktek](https://academy.dqlab.id/main/livecode/161/304/1376)
